auto_get_task
A program for automatically receiving tasks. The script looks for a button on the screen and presses it with a delay of 1 second until the button disappears, which means that the task has been received. Powered by PyAutoGui. When a task is received, a sound is played and the script stops working.